Algorithm for PA 5

+++ Onyuka-PA5.py
@@ -75,7 +75,7 @@ def average_magnitude (earthquakes):
 def num_of_earthquakes_in_distance(earthquakes,lat1,lon1,user_distance ):
     count=0
     for i in range(len(earthquakes)):
-        distance =(math.acos(math.sin(lat1) * math.sin(int(earthquakes[1])) + math.cos(lat1) * math.cos(earthquakes[1]) * math.cos(lon1 - earthquakes[2])) * 6371)
+        distance =float(math.acos(math.sin(lat1) * math.sin(int(earthquakes[1])) + math.cos(lat1) * math.cos(earthquakes[1]) * math.cos(lon1 - earthquakes[2])) * 6371)
         if distance <= user_distance:
             count+=1
     return count
